requirements.txt
absl-py==2.3.0
agno==1.6.3
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
argbind==0.3.9
asttokens==3.0.0
attrs==25.3.0
audioread==3.0.1
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-config-file==0.6.0
configobj==5.0.9
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.16
einops==0.8.1
elevenlabs==2.5.0
executing==2.2.0
fastapi==0.115.13
ffmpy==0.6.0
filelock==3.18.0
fire==0.7.0
flatten-dict==0.4.2
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
google-genai==1.21.1
gotrue==2.12.0
gradio==5.35.0
gradio_client==1.10.4
groovy==0.1.2
grpcio==1.73.1
gunicorn==22.0.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.1
hyperframe==6.1.0
idna==3.10
importlib_resources==6.5.2
iniconfig==2.1.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
julius==0.2.7
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
Markdown==3.8.2
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.1
multidict==6.5.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
orjson==3.10.18
packaging==25.0
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
pooch==1.8.2
postgrest==1.0.2
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==3.19.6
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pylance==0.30.0
pyloudnorm==0.1.1
pyparsing==3.2.3
pystoi==0.4.1
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
randomname==0.2.1
realtime==2.4.3
requests==2.32.4
rich==14.0.0
rsa==4.9.1
ruff==0.12.1
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
semantic-version==2.10.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
stack-data==0.6.3
starlette==0.46.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.3
supafunc==0.9.4
sympy==1.13.1
tenacity==8.5.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
threadpoolctl==3.6.0
tomli==2.2.1
tomlkit==0.13.3
torch==2.6.0
torch-stoi==0.2.3
torchaudio==2.6.0
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
wcwidth==0.2.13
websockets==14.2
Werkzeug==3.1.3
yarl==1.20.1
